Update criteria for test generation
If a "known option" was encountered while looking in the hashmap
containing pre-defined option definitions, a testcase to check the
validity of the result of that option is produced. In case no such
option was encountered, testcases to verify the correct usage message for
all of the valid options are produced when they are used incorrectly.