Shift to using clang for building in FreeBSD
The cause of segmentation fault encountered while execution was that the
tool was built using gcc and the attempt was being made to link it with
the boost libraries (installed via ports) which are built using clang.
Using clang to build the tool produces the expected behavior in FreeBSD.
Thanks to @asomers for clarifying the issue!

Reference: https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=221882