Update criteria for adding variable "usage_output"
As noted in afe9f92961f2d0915d63a9ae0d415d1abeff7978, the assumption -
if the usage message is consistent across "atleast" two options, then it
should be same for all of them - is false and tested to fail for
mkdir(1). This issue is fixed by finding two matching usage messages
from a list of 3 (by making 3 comparisons). The value 3 is chosen as it
seems good enough to come up with the decision for using "usage_output"
variable, at a cost of only 3 comparisons. A more rigorous approach will
use more than 3 options, but then the number of comparisons will
increase by a notable factor. If a case arises in future for which the
value 3 is not sufficient, then it will be incremented.

Fixes #15