[WIP] First attempt at terminating a running thread at will
To achieve this, we first detach the thread (which I am assuming places
the thread in a different thread group), and then send it a SIGINT.
The expected behavior of this is that the thread represented by its
handle will be killed, and the execution will continue normally.
However, the observed behavior for this is that pthread_kill() returns
with an exit status 22, and the entire program terminates. A plausible
reasoning behind this is that the thread handle returned by
"native_handle()" is of the currently running process itself. Another
reason for this can be that the detached thread is still in the thread
group of the currently running parent process, and when the thread is
killed, the parent dies along with it.

An alternative approach for this (which I earlier used) is that we don't
perform any cleanup for the thread corresponding to the failed test. In
that case, the next scheduled run for generating test for stdbuf(1) (which
succeeds normally) will start and then fail too <- "I haven't been able to
figure out the reason behind this behavior".

Other introduced changes:
  * Rename "Exec()" => "Execute()".
  * Place "Execute()" in utils.cpp as it is a utility function.
  * Manually remove the groff scripts (while fetching them in
    src/fetch_groff.sh) which cause an unexpected termination. The only
    reason behind this is to get a clean run, and will not be required
    once the fix described above works.
  * Add functionality to place the groff scripts corresponding to the
    failed tests in a separate directory named "failed_groff".