Add functionality to kill (child) process when stuck on a blocking read
Instead of spawning a thread for each utility, we create a (child) shell
process for executing the shell command and mux its file descriptors via
select() with an introduced timeout for finishing execution.

When pclose() is called on the stream returned by popen(), it waits
indefinitely for the created shell process to terminate in cass where
the shell command waits for the user input via a blocking read (e.g.
passwd(1)). Hence, we define a custom function which alongside returning
the FILE* stream (as with popen()) also returns the pid of the newly
created (child) shell process, which can later be used for signalling.

For reference, the threaded implementation is available on ~
Branch: "threaded"
Commit: 6409d3f27e47cc108e37f614c64d5574b6d3e970